requirements.txt
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
PyYAML==6.0
scikit_learn==1.2.2
setuptools==45.2.0
torch==2.0.0
tqdm==4.42.1
